requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.16,<3.17
django-cors-headers>=4.9,<5

